MAJOR UPDATE DATA, DATA PASSING, AND POOL GEN
The match data class is now parcelable. The tournament class has a genpools
that functions, along with a tester class for that. When you enter contestants,
the pools are updated as well. And the pools will generate the correct amount
of grid spaces.

Next, I will work on passing the tournament to the pool activity, so that we
can display the names of the contestants.